refactor:라우팅 작업을 위한 폴더 구조 변경
