+++ src/page-layout/FolderPageLayout/FolderPageLayout.css
@@ -1,0 +1,28 @@
+.FolderPageLayout {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 100%;
+  width: 100%;
+  row-gap: 2rem;
+  padding-bottom: 2rem;
+
+  @media (min-width: 768px) {
+    row-gap: 4rem;
+    padding-bottom: 6rem;
+  }
+}
+
+.FolderPageLayout-items {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  max-width: 112.4rem;
+  row-gap: 3.2rem;
+  padding: 0 3.2rem;
+
+  @media (min-width: 768px) {
+    row-gap: 4rem;
+  }
+}
