+++ src/pages/SharedPage.jsx
@@ -1,10 +1,10 @@
-import { SharedPageLayout } from "../SharedPageLayout/SharedPageLayout";
+import { SharedPageLayout } from "../page-layout/SharedPageLayout/SharedPageLayout";
 import { useGetFolder } from "../data-access/useGetFolder";
-import { Layout } from "../feature/Layout/Layout";
-import { CardList } from "../ui/CardList/CardList";
-import { FolderInfo } from "../ui/FolderInfo/FolderInfo";
-import { ReadOnlyCard } from "../ui/ReadOnlyCard/RedOnlyCard";
-import { SearchBar } from "../ui/SearchBar/SearchBar";
+import { Layout } from "../sharing/Layout/Layout";
+import { CardList } from "../sharing/card/CardList/CardList";
+import { FolderInfo } from "../shared-page-ui/FolderInfo/FolderInfo";
+import { ReadOnlyCard } from "../sharing/card/ReadOnlyCard/RedOnlyCard";
+import { SearchBar } from "../sharing/SearchBar/SearchBar";
 
 export const SharedPage = () => {
   const { data } = useGetFolder();
