+++ src/shared-page-ui/FolderInfo/FolderInfo.css
