+++ src/shared-page-ui/FolderInfo/FolderInfo.js
