+++ src/sharing/Footer/constant.js
