+++ src/sharing/NavigationBar/NavigationBar.css
