+++ src/sharing/NavigationBar/constant.js
