+++ src/sharing/SearchBar/SearchBar.css
