+++ src/sharing/SearchBar/SearchBar.js
