+++ src/sharing/SearchBar/constant.js
