+++ src/sharing/card/CardContent/CardContent.js
